requirements.txt
python-dotenv
discord
instaloader
asyncio
PyYAML==6.0.1